Added a manifest template for research_timer. research_timer is a C++ program, so running the program needs extra dependencies, e.g., libstdc++.so.6. To make graphene use these dependencies, first the file system needs to be cherry-picked into the guest file system of the library OS. Two file system mount points are added to the manifest.
Secondly, the environment variable LD_LIBRARY_PATH needs to include the
directories where the dynamic linker (i.e., ld) can find these
dependencies (e.g., /usr/lib/x86_64-linux-gnu for libstdc++.so.6).

Finally, to run in an SGX enclave, the paths of these dependencies need
to be added as *trusted files* in the manifest.

------

In addition, a patch is added to fix the compilation of a C++ source file
in Graphene's LibOS/shim/test/native directory. To apply the patch, simply
copy graphene-tests-cpp.patch into the Graphene local repository, and run:

    git apply graphene-tests-cpp.patch